refactor(update): keep files under the 500-line structure limit
The nightly-channel commit tipped two files over the repo's 500-line gate.
Extract the channel selection methods (channel/set_channel/pending switch) from
services/releases/mod.rs into a sibling `channel.rs`, and move the channel
decision test next to `evaluate_parsed` in `verify.rs` (updater/ is already at
the 8-entries-per-dir limit, so no new file there). Pure code motion — no
behaviour change.
